pyproject.toml
[project]
name = "dawn"
version = "0.1.0"
description = ""
authors = [
    {name = "Danny McGrory",email = "dannyalaska@mac.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.49.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "boto3 (>=1.40.32,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "xlrd (>=2.0.2,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "sentence-transformers (>=5.1.1,<6.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "openai (>=2.1.0,<3.0.0)",
    "sqlglot (>=25.13.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "ruff (>=0.13.0,<0.14.0)",
    "black (>=25.1.0,<26.0.0)",
    "mypy (>=1.18.1,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "types-redis (>=4.6.0.20241004,<5.0.0.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E","F","I","B","UP","SIM"]
lint.ignore = ["E501"]
extend-exclude = ["infra/"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing"
testpaths = ["app/tests"]
